Move simple_test_runner out of plugins and into core library.
Now if you do `list` with no plugins installed, you actually see 0 plugins.
